+++ ror-tools-core/src/main/scala/tech/beshu/ror/tools/core/patches/Es70xPatch.scala
@@ -21,6 +21,9 @@ import tech.beshu.ror.tools.core.patches.base.SimpleEsPatch
 import tech.beshu.ror.tools.core.patches.internal.RorPluginDirectory
 import tech.beshu.ror.tools.core.patches.internal.filePatchers.*
 import tech.beshu.ror.tools.core.patches.internal.modifiers.bytecodeJars.*
+import tech.beshu.ror.tools.core.patches.internal.modifiers.bytecodeJars.authentication.{DummyAuthenticationInAuthenticationServiceAuthenticator, GetAuthenticationFromHeaderWhenMissingInTransient}
+import tech.beshu.ror.tools.core.patches.internal.modifiers.bytecodeJars.authorization.{DummyAuthorizeInAuthorizationService, MockAuthorizationInfoInAuthorizationService}
+import tech.beshu.ror.tools.core.patches.internal.modifiers.bytecodeJars.permissions.AlwaysGrantApplicationPermission
 
 import scala.language.postfixOps
