+++ ror-tools-core/src/main/scala/tech/beshu/ror/tools/core/patches/Es83xPatch.scala
@@ -22,6 +22,9 @@ import tech.beshu.ror.tools.core.patches.base.TransportNetty4AwareEsPatch
 import tech.beshu.ror.tools.core.patches.internal.RorPluginDirectory
 import tech.beshu.ror.tools.core.patches.internal.filePatchers.*
 import tech.beshu.ror.tools.core.patches.internal.modifiers.bytecodeJars.*
+import tech.beshu.ror.tools.core.patches.internal.modifiers.bytecodeJars.authentication.DummyAuthenticationInAuthenticationChain
+import tech.beshu.ror.tools.core.patches.internal.modifiers.bytecodeJars.authorization.{DummyAuthorizeInAuthorizationService, MockAuthorizationInfoDuringRetrievingUserPrivilegesInAuthorizationService, MockAuthorizationInfoInSecurityContext}
+import tech.beshu.ror.tools.core.patches.internal.modifiers.bytecodeJars.permissions.{AlwaysGrantApplicationPermission, ModifyBootstrapPolicyUtilClass}
 import tech.beshu.ror.tools.core.patches.internal.modifiers.securityPolicyFiles.AddAdditionalPermissions
 import tech.beshu.ror.tools.core.patches.internal.modifiers.securityPolicyFiles.AddAdditionalPermissions.*
